More TagBakery changes
